+++ layout/explore/explore-areas-of-interest/index.js
@@ -2,11 +2,8 @@ import { connect } from 'react-redux';
 import {
   setSidebarSubsection,
   setSelectedItem,
-  toggleArea,
+  setAreaOfInterest,
 } from 'layout/explore/actions';
-
-// selectors
-import { getAreaIds } from './selectors';
 
 // component
 import ExploreAreasOfInterest from './component';
@@ -14,11 +11,11 @@ import ExploreAreasOfInterest from './component';
 export default connect(
   (state) => ({
     token: state.user.token,
-    areasOnMap: getAreaIds(state),
+    aoi: state.explore.map.aoi,
   }),
   {
     setSidebarSubsection,
     setSelectedItem,
-    toggleArea,
+    setAreaOfInterest,
   },
 )(ExploreAreasOfInterest);