+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [2.17.8] - TBD
 ### Added
+- Explore: added `aoi` param to URL to allow loading areas of interest from the URL.
 - pagination for collections coming from API.
 - call action in areas of interest to go to all-data section.
 - Explore: added active state for prominent buttons in new area form.
@@ -14,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - widget-editor: WRI's colour scheme.
 
 ### Changed
+- Explore map now displays one area of interest at the same time. 
 - In MyRW, collections without resources are hide depending on the tab the user is.
 - updates icons for widgets and datasets in collections table.
 - moved collections away from redux.